Add Dockerfile for ASP.Net 7.0 with required utilities
This commit adds a new Dockerfile for the ASP.NET 7.0 base image. It also includes the installation of necessary utilities such as curl, unzip and openjdk-17-jdk. The image is prepared with apt-get updates, locale settings and environment variables. This setup is needed for effectively building and running the ASP.NET 7.0 applications.